requirements.txt
google-play-scraper>=1.2.7
requests>=2.31.0
tavily-python>=0.3.0
app-store-scraper>=0.3.5
